unified user and tasks routes under api.py
